+++ demos/58-shape-reflecting-rays/index.js
@@ -27,6 +27,8 @@
         GUP
       )
     );
+    // Disable automatically added handle lines
+    pb.drawConfig.drawHandleLines = false;
     // Let's set up some colors.
     pb.drawConfig.polygon.color = "rgba(128,128,128,0.5)";
     pb.drawConfig.polygon.lineWidth = 2;
@@ -41,27 +43,25 @@
 
     // Array<Polygon | Circle | Ellipse>
     var shapes = [];
+    var mainRay = new Vector(new Vertex(), new Vertex(250, 250).rotate(Math.random() * Math.PI));
     // Array<Vector>
     var rays = [];
     var ellipseHelper;
     var cicleSectorHelper;
     var ellipseSectorHelper;
-    rays.push(new Vector(new Vertex(), new Vertex(250, 250).rotate(Math.random() * Math.PI)));
+    rays.push(mainRay);
 
     // Create a config: we want to have control about the arrow head size in this demo
     var config = {
-      pointCount: params.getNumber("pointCount", 4)
+      animate: params.getBoolean("animate", false)
     };
 
     // +---------------------------------------------------------------------------------
     // | Global vars
     // +-------------------------------
     var viewport = pb.viewport();
-    // Must be in clockwise order!
-    var polygon = null;
 
     var postDraw = function (draw, fill) {
-      // draw.circle({ x: 0, y: 0 }, 50.0, "red", 1.0);
       var reflectedRaysByShapes = calculateAllReflections();
       reflectedRaysByShapes.forEach(function (reflectedRays) {
         reflectedRays.forEach(function (ray) {
@@ -71,14 +71,6 @@
       ellipseHelper.drawHandleLines(draw, fill);
       cicleSectorHelper.drawHandleLines(draw, fill);
       ellipseSectorHelper.drawHandleLines(draw, fill);
-
-      // Draw barycenter of polygon?
-      // shapes.forEach(function (shape) {
-      //   if (shape instanceof Polygon) {
-      //     console.log("centroid", shape.getCentroid());
-      //     draw.diamondHandle(shape.getCentroid(), 5, "orange");
-      //   }
-      // });
     };
 
     /**
@@ -107,7 +99,7 @@
      * @returns
      */
     var findReflectedRay = function (shape, ray) {
-      var reflectedRay = null; // ray.perp().moveTo(ray.b);
+      var reflectedRay = null;
 
       // Find intersection with min distance
       if (
@@ -156,15 +148,33 @@
     // | Just rebuilds the pattern on changes.
     // +-------------------------------
     var rebuildShape = function () {
-      // Create a new randomized polygon.
-      polygon = createRandomizedPolygon(config.pointCount, viewport, true); // createClockwise=true
-
-      // Re-add the new polygon to trigger redraw.
       pb.removeAll(false, false); // Don't trigger redraw
+
+      // Create a new randomized polygon.
+      var polygon = createRandomizedPolygon(4, viewport, true); // createClockwise=true
+      polygon.scale(0.3, polygon.getCentroid());
 
       // Create circle and ellpise
       var circle = new Circle(new Vertex(-25, -15), 90.0);
       var ellipse = new VEllipse(new Vertex(25, 15), new Vertex(150, 200), -Math.PI * 0.3);
+      var circleSector = randomCircleSector(viewport);
+      var ellipseSector = randomEllipseSector(viewport);
+
+      // Create a new randomized Bézier curve.
+      var tmpPolygon = createRandomizedPolygon(4, viewport, true); // createClockwise=true
+      tmpPolygon.scale(0.3, polygon.getCentroid());
+      var bezierPath = BezierPath.fromCurve(
+        new CubicBezierCurve(tmpPolygon.vertices[0], tmpPolygon.vertices[1], tmpPolygon.vertices[2], tmpPolygon.vertices[3])
+      );
+
+      shapes = [polygon, circle, ellipse, circleSector, ellipseSector, bezierPath];
+      // Align all shapes on a circle :)
+      var alignCircle = new Circle(new Vertex(), viewport.getMinDimension() * 0.333);
+      shapes.forEach(function (shape, i) {
+        var newPosition = alignCircle.vertAt((i * Math.PI * 2) / shapes.length);
+        // console.log("shape ", i, typeof shape);
+        shape.move(newPosition);
+      });
 
       // We want to change the ellipse's radii and rotation by dragging points around
       var ellipseRotationControlPoint = ellipse.vertAt(0.0).scale(1.2, ellipse.center);
@@ -173,8 +183,6 @@
       }
       ellipseHelper = new VEllipseHelper(ellipse, ellipseRotationControlPoint);
 
-      // Create circle sector
-      var circleSector = randomCircleSector(viewport);
       // Further: add a circle sector helper to edit angles and radius manually (mouse or touch)
       var controlPointA = circleSector.circle.vertAt(circleSector.startAngle);
       var controlPointB = circleSector.circle.vertAt(circleSector.endAngle);
@@ -183,8 +191,6 @@
       }
       cicleSectorHelper = new CircleSectorHelper(circleSector, controlPointA, controlPointB, pb);
 
-      // Create an ellipse helper
-      var ellipseSector = randomEllipseSector(viewport);
       // We want to change the ellipse's radii and rotation by dragging points around
       var startControlPoint = ellipseSector.ellipse.vertAt(ellipseSector.startAngle);
       var endControlPoint = ellipseSector.ellipse.vertAt(ellipseSector.endAngle);
@@ -196,22 +202,42 @@
       }
       ellipseSectorHelper = new VEllipseSectorHelper(ellipseSector, startControlPoint, endControlPoint, rotationControlPoint);
 
-      shapes = [
-        polygon,
-        circle,
-        ellipse,
-        ellipseRotationControlPoint,
-        circleSector,
-        controlPointA,
-        controlPointB,
-        ellipseSector,
-        startControlPoint,
-        endControlPoint,
-        rotationControlPoint
-      ];
       pb.add(shapes, false);
+      pb.add(
+        [ellipseRotationControlPoint, controlPointA, controlPointB, startControlPoint, endControlPoint, rotationControlPoint],
+        false
+      );
       pb.add(rays, true); // trigger redraw
     };
+
+    // +---------------------------------------------------------------------------------
+    // | Render next animation step.
+    // +-------------------------------
+    function animateStep(time) {
+      var animationCircle = new Circle(new Vertex(), viewport.getMinDimension() * 0.333);
+      mainRay.b.set(animationCircle.vertAt(time / 5000));
+      pb.redraw();
+      if (isAnimationRunning) {
+        globalThis.requestAnimationFrame(animateStep);
+      }
+    }
+
+    // +---------------------------------------------------------------------------------
+    // | Toggle animation of main ray.
+    // +-------------------------------
+    var isAnimationRunning = false;
+    function toggleAnimation() {
+      if (config.animate) {
+        if (!isAnimationRunning) {
+          isAnimationRunning = true;
+          animateStep(0);
+        }
+      } else {
+        if (isAnimationRunning) {
+          isAnimationRunning = false;
+        }
+      }
+    }
 
     // +---------------------------------------------------------------------------------
     // | Create a GUI.
@@ -219,11 +245,12 @@
     {
       var gui = pb.createGUI();
       // prettier-ignore
-      gui.add(config, "pointCount").min(3).max(32).step(1).name("pointCount").title("Number of polygon vertices")
-      .onChange( function() { rebuildShape(); });
+      gui.add(config, "animate").name("animate").title("Animate the ray?")
+      .onChange( function() { toggleAnimation(); });
     }
 
     pb.config.postDraw = postDraw;
     rebuildShape();
+    toggleAnimation();
   });
 })(globalThis);