+++ src/ts/CircleSector.ts
@@ -8,7 +8,8 @@
  * @modified 2024-03-08 Added the `containsAngle` method.
  * @modified 2024-03-09 Added the `circleSectorIntersection` method to find coherent sector intersections..
  * @modified 2024-03-09 Added the `angleAt` method to determine any angle at some ratio.
- * @modified 2025-04-02 Adding `CircleSector.lineIntersections` and `CircleSector.lineIntersectionTangents` and implementing `Intersectable`.
+ * @modified 2025-04-02 Adding the `CircleSector.lineIntersections` and `CircleSector.lineIntersectionTangents` and implementing `Intersectable`.
+ * @modified 2025-04-09 Adding the `CircleSector.move()` method.
  * @version  1.2.0
  **/
 
@@ -89,6 +90,20 @@ export class CircleSector implements Intersectable, SVGSerializable {
     this.circle = circle;
     this.startAngle = startAngle;
     this.endAngle = endAngle;
+  }
+
+  /**
+   * Move the circle sector by the given amount.
+   *
+   * @method move
+   * @param {XYCoords} amount - The amount to move.
+   * @instance
+   * @memberof CircleSector
+   * @return {CircleSector} this for chaining
+   **/
+  move(amount: XYCoords): CircleSector {
+    this.circle.move(amount);
+    return this;
   }
 
   /**